requirements.txt
torch>=2.0.1
torchvision>=0.15.2
torchaudio>=2.0.2
diffusers>=0.21.4
transformers>=4.31.0
accelerate>=0.21.0
pillow>=10.0.0
numpy>=1.24.3
safetensors>=0.3.3
xformers>=0.0.20